py/pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.76.0",
    "python-dotenv>=1.2.2",
]
